item rarity in stroke style
